Move benchmark reporting into module; remove global variable
